backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
jinja2
loguru